os/filestore: move FileStore to os/filestore/*
