+++ src/os/filestore/FileStore.cc
@@ -1,4 +1,4 @@
-// -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*- 
+// -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
 // vim: ts=8 sw=2 smarttab
 /*
  * Ceph - scalable distributed file system
@@ -8,9 +8,9 @@
  *
  * This is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
- * License version 2.1, as published by the Free Software 
+ * License version 2.1, as published by the Free Software
  * Foundation.  See file COPYING.
- * 
+ *
  */
 #include "include/int_types.h"
 
@@ -150,7 +150,7 @@ ostream& operator<<(ostream& out, const FileStore::OpSequencer& s)
   return out << *s.parent;
 }
 
-int FileStore::get_cdir(coll_t cid, char *s, int len) 
+int FileStore::get_cdir(coll_t cid, char *s, int len)
 {
   const string &cid_str(cid.to_str());
   return snprintf(s, len, "%s/current/%s", basedir.c_str(), cid_str.c_str());
@@ -379,7 +379,7 @@ int FileStore::lfn_link(coll_t c, coll_t newcid, const ghobject_t& o, const ghob
     }
     if (!exist)
       return -ENOENT;
-  
+
     RWLock::WLocker l2((index_new.index)->access_lock);
 
     r = index_new->lookup(newoid, &path_new, &exist);
@@ -434,7 +434,7 @@ int FileStore::lfn_link(coll_t c, coll_t newcid, const ghobject_t& o, const ghob
       assert(!m_filestore_fail_eio || r != -EIO);
       return r;
     }
-  }    
+  }
   return 0;
 }
 
@@ -511,7 +511,7 @@ FileStore::FileStore(const std::string &base, const std::string &jdev, osflagbit
   backend(NULL),
   index_manager(do_update),
   lock("FileStore::lock"),
-  force_sync(false), 
+  force_sync(false),
   sync_entry_timeo_lock("sync_entry_timeo_lock"),
   timer(g_ceph_context, sync_entry_timeo_lock),
   stop(false), sync_thread(this),
@@ -657,7 +657,7 @@ void FileStore::collect_metadata(map<string,string> *pm)
   char partition_path[PATH_MAX];
   char dev_node[PATH_MAX];
   int rc = 0;
-  
+
   (*pm)["filestore_backend"] = backend->get_name();
   ostringstream ss;
   ss << "0x" << std::hex << m_fs_type << std::dec;
@@ -898,13 +898,13 @@ int FileStore::mkfs()
 	ret = backend->create_checkpoint(s, NULL);
 	VOID_TEMP_FAILURE_RETRY(::close(current_fd));
 	if (ret < 0 && ret != -EEXIST) {
-	  VOID_TEMP_FAILURE_RETRY(::close(fd));  
+	  VOID_TEMP_FAILURE_RETRY(::close(fd));
 	  derr << "mkfs: failed to create snap_1: " << cpp_strerror(ret) << dendl;
 	  goto close_fsid_fd;
 	}
       }
     }
-    VOID_TEMP_FAILURE_RETRY(::close(fd));  
+    VOID_TEMP_FAILURE_RETRY(::close(fd));
   }
   ret = KeyValueDB::test_init(superblock.omap_backend, omap_dir);
   if (ret < 0) {
@@ -1103,7 +1103,7 @@ int FileStore::_sanity_check_fs()
       (int)m_filestore_journal_parallel +
       (int)m_filestore_journal_trailing) > 1) {
     dout(0) << "mount ERROR: more than one of filestore journal {writeahead,parallel,trailing} enabled" << dendl;
-    cerr << TEXT_RED 
+    cerr << TEXT_RED
 	 << " ** WARNING: more than one of 'filestore journal {writeahead,parallel,trailing}'\n"
 	 << "             is enabled in ceph.conf.  You must choose a single journal mode."
 	 << TEXT_NORMAL << std::endl;
@@ -1113,7 +1113,7 @@ int FileStore::_sanity_check_fs()
   if (!backend->can_checkpoint()) {
     if (!journal || !m_filestore_journal_writeahead) {
       dout(0) << "mount WARNING: no btrfs, and no journal in writeahead mode; data may be lost" << dendl;
-      cerr << TEXT_RED 
+      cerr << TEXT_RED
 	   << " ** WARNING: no btrfs AND (no journal OR journal not in writeahead mode)\n"
 	   << "             For non-btrfs volumes, a writeahead journal is required to\n"
 	   << "             maintain on-disk consistency in the event of a crash.  Your conf\n"
@@ -1266,7 +1266,7 @@ int FileStore::mount()
   CompatSet supported_compat_set = get_fs_supported_compat_set();
 
   dout(5) << "basedir " << basedir << " journal " << journalpath << dendl;
-  
+
   // make sure global base dir exists
   if (::access(basedir.c_str(), R_OK | W_OK)) {
     ret = -errno;
@@ -1310,7 +1310,7 @@ int FileStore::mount()
   } else if (ret == 0) {
     if (do_update || (int)version_stamp < g_conf->filestore_update_to) {
       derr << "FileStore::mount : stale version stamp detected: "
-	   << version_stamp 
+	   << version_stamp
 	   << ". Proceeding, do_update "
 	   << "is set, performing disk format upgrade."
 	   << dendl;
@@ -1580,7 +1580,7 @@ int FileStore::mount()
       Index index;
       ret = get_index(*i, &index);
       if (ret < 0) {
-	derr << "Unable to mount index " << *i 
+	derr << "Unable to mount index " << *i
 	     << " with error: " << ret << dendl;
 	goto close_current_fd;
       }
@@ -1705,10 +1705,10 @@ void FileStore::init_temp_collections()
   }
 }
 
-int FileStore::umount() 
+int FileStore::umount()
 {
   dout(5) << "umount " << basedir << dendl;
-  
+
   flush();
   sync();
   do_force_sync();
@@ -1831,7 +1831,7 @@ void FileStore::op_queue_reserve_throttle(Op *o, ThreadPool::TPHandle *handle)
 
   if (handle)
     handle->suspend_tp_timeout();
-  if (throttle_ops.should_wait(1) || 
+  if (throttle_ops.should_wait(1) ||
     (throttle_bytes.get_current()      // let single large ops through!
     && throttle_bytes.should_wait(o->bytes))) {
     dout(2) << "waiting " << throttle_ops.get_current() + 1 << " > " << max_ops << " ops || "
@@ -1881,7 +1881,7 @@ void FileStore::_finish_op(OpSequencer *osr)
 {
   list<Context*> to_queue;
   Op *o = osr->dequeue(&to_queue);
-  
+
   utime_t lat = ceph_clock_now(g_ceph_context);
   lat -= o->start;
 
@@ -1971,14 +1971,14 @@ int FileStore::queue_transactions(Sequencer *posr, list<Transaction*> &tls,
 
     if (m_filestore_journal_parallel) {
       dout(5) << "queue_transactions (parallel) " << o->op << " " << o->tls << dendl;
-      
+
       _op_journal_transactions(tbl, orig_len, o->op, ondisk, osd_op);
-      
+
       // queue inside submit_manager op submission lock
       queue_op(osr, o);
     } else if (m_filestore_journal_writeahead) {
       dout(5) << "queue_transactions (writeahead) " << o->op << " " << o->tls << dendl;
-      
+
       osr->queue_journal(o->op);
 
       _op_journal_transactions(tbl, orig_len, o->op,
@@ -2030,7 +2030,7 @@ int FileStore::queue_transactions(Sequencer *posr, list<Transaction*> &tls,
 
   apply_manager.op_apply_start(op);
   int r = do_transactions(tls, op);
-    
+
   if (r >= 0) {
     _op_journal_transactions(tbl, orig_len, op, ondisk, osd_op);
   } else {
@@ -2090,7 +2090,7 @@ int FileStore::_do_transactions(
     if (handle)
       handle->reset_tp_timeout();
   }
-  
+
   return r;
 }
 
@@ -2187,7 +2187,7 @@ void FileStore::_set_replay_guard(coll_t cid,
   }
   _set_replay_guard(fd, spos, 0, in_progress);
   VOID_TEMP_FAILURE_RETRY(::close(fd));
-} 
+}
 
 
 void FileStore::_set_replay_guard(int fd,
@@ -2243,7 +2243,7 @@ void FileStore::_close_replay_guard(coll_t cid,
   }
   _close_replay_guard(fd, spos);
   VOID_TEMP_FAILURE_RETRY(::close(fd));
-} 
+}
 
 void FileStore::_close_replay_guard(int fd, const SequencerPosition& spos)
 {
@@ -2363,7 +2363,7 @@ unsigned FileStore::_do_transaction(
 #endif
 
   Transaction::iterator i = t.begin();
-  
+
   SequencerPosition spos(op_seq, trans_num, 0);
   while (i.have_op()) {
     if (handle)
@@ -2388,7 +2388,7 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, touch_exit, r);
       }
       break;
-      
+
     case Transaction::OP_WRITE:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2405,7 +2405,7 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, write_exit, r);
       }
       break;
-      
+
     case Transaction::OP_ZERO:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2419,13 +2419,13 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, zero_exit, r);
       }
       break;
-      
+
     case Transaction::OP_TRIMCACHE:
       {
 	// deprecated, no-op
       }
       break;
-      
+
     case Transaction::OP_TRUNCATE:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2438,7 +2438,7 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, truncate_exit, r);
       }
       break;
-      
+
     case Transaction::OP_REMOVE:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2450,7 +2450,7 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, remove_exit, r);
       }
       break;
-      
+
     case Transaction::OP_SETATTR:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2471,7 +2471,7 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, setattr_exit, r);
       }
       break;
-      
+
     case Transaction::OP_SETATTRS:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2512,7 +2512,7 @@ unsigned FileStore::_do_transaction(
         tracepoint(objectstore, rmattrs_exit, r);
       }
       break;
-      
+
     case Transaction::OP_CLONE:
       {
         coll_t cid = i.get_cid(op->cid);
@@ -2892,7 +2892,7 @@ bool FileStore::exists(coll_t cid, const ghobject_t& oid)
   tracepoint(objectstore, exists_exit, retval);
   return retval;
 }
-  
+
 int FileStore::stat(
   coll_t cid, const ghobject_t& oid, struct stat *st, bool allow_eio)
 {
@@ -3125,7 +3125,7 @@ int FileStore::fiemap(coll_t cid, const ghobject_t& oid,
     dout(10) << "read couldn't open " << cid << "/" << oid << ": " << cpp_strerror(r) << dendl;
     goto done;
   }
-  
+
   if (backend->has_seek_data_hole()) {
     dout(15) << "seek_data/seek_hole " << cid << "/" << oid << " " << offset << "~" << len << dendl;
     r = _do_seek_hole_data(**fd, offset, len, &exomap);
@@ -3148,7 +3148,7 @@ done:
 
 
 int FileStore::_remove(coll_t cid, const ghobject_t& oid,
-		       const SequencerPosition &spos) 
+		       const SequencerPosition &spos)
 {
   dout(15) << "remove " << cid << "/" << oid << dendl;
   int r = lfn_unlink(cid, oid, spos);
@@ -3197,7 +3197,7 @@ int FileStore::_write(coll_t cid, const ghobject_t& oid,
 	    << cpp_strerror(r) << dendl;
     goto out;
   }
-    
+
   // seek
   actual = ::lseek64(**fd, offset, SEEK_SET);
   if (actual < 0) {
@@ -3579,7 +3579,7 @@ int FileStore::_clone_range(coll_t cid, const ghobject_t& oldoid, const ghobject
 
 class SyncEntryTimeout : public Context {
 public:
-  SyncEntryTimeout(int commit_timeo) 
+  SyncEntryTimeout(int commit_timeo)
     : m_commit_timeo(commit_timeo)
   {
   }
@@ -3625,7 +3625,7 @@ void FileStore::sync_entry()
       if (woke < min_interval) {
 	utime_t t = min_interval;
 	t -= woke;
-	dout(20) << "sync_entry waiting for another " << t 
+	dout(20) << "sync_entry waiting for another " << t
 		 << " to reach min interval " << min_interval << dendl;
 	sync_cond.WaitInterval(g_ceph_context, lock, t);
       }
@@ -3635,7 +3635,7 @@ void FileStore::sync_entry()
   again:
     fin.swap(sync_waiters);
     lock.Unlock();
-    
+
     op_tp.pause();
     if (apply_manager.commit_start()) {
       utime_t start = ceph_clock_now(g_ceph_context);
@@ -3709,7 +3709,7 @@ void FileStore::sync_entry()
 	  assert(0 == "error during fsync of op_seq");
 	}
       }
-      
+
       utime_t done = ceph_clock_now(g_ceph_context);
       utime_t lat = done - start;
       utime_t dur = done - startwait;
@@ -3747,7 +3747,7 @@ void FileStore::sync_entry()
     } else {
       op_tp.unpause();
     }
-    
+
     lock.Lock();
     finish_contexts(g_ceph_context, fin, 0);
     fin.clear();
@@ -3835,7 +3835,7 @@ void FileStore::flush()
       cond.Wait(lock);
     assert(0);
   }
- 
+
   if (m_filestore_journal_writeahead) {
     if (journal)
       journal->flush();
@@ -4180,7 +4180,7 @@ int FileStore::_setattrs(coll_t cid, const ghobject_t& oid, map<string,bufferptr
   incomplete_inline = (r == -E2BIG);
   assert(!m_filestore_fail_eio || r != -EIO);
   dout(15) << "setattrs " << cid << "/" << oid
-    	   << (incomplete_inline ? " (incomplete_inline, forcing omap)" : "")
+	   << (incomplete_inline ? " (incomplete_inline, forcing omap)" : "")
 	   << dendl;
 
   for (map<string,bufferptr>::iterator p = aset.begin();
@@ -4549,7 +4549,7 @@ int FileStore::collection_version_current(coll_t c, uint32_t *version)
   *version = index->collection_version();
   if (*version == target_version)
     return 1;
-  else 
+  else
     return 0;
 }
 
@@ -4627,7 +4627,7 @@ int FileStore::list_collections(vector<coll_t>& ls, bool include_temp)
   return r;
 }
 
-int FileStore::collection_stat(coll_t c, struct stat *st) 
+int FileStore::collection_stat(coll_t c, struct stat *st)
 {
   tracepoint(objectstore, collection_stat_enter, c.c_str());
   char fn[PATH_MAX];
@@ -4642,7 +4642,7 @@ int FileStore::collection_stat(coll_t c, struct stat *st)
   return r;
 }
 
-bool FileStore::collection_exists(coll_t c) 
+bool FileStore::collection_exists(coll_t c)
 {
   tracepoint(objectstore, collection_exists_enter, c.c_str());
   struct stat st;
@@ -4651,8 +4651,8 @@ bool FileStore::collection_exists(coll_t c)
   return ret;
 }
 
-bool FileStore::collection_empty(coll_t c) 
-{  
+bool FileStore::collection_empty(coll_t c)
+{
   tracepoint(objectstore, collection_empty_enter, c.c_str());
   dout(15) << "collection_empty " << c << dendl;
   Index index;
@@ -4910,7 +4910,7 @@ ObjectMap::ObjectMapIterator FileStore::get_omap_iterator(coll_t c,
   if (r < 0) {
     dout(10) << __func__ << " " << c << "/" << hoid << " = 0 "
 	     << "(get_index failed with " << cpp_strerror(r) << ")" << dendl;
-    return ObjectMap::ObjectMapIterator(); 
+    return ObjectMap::ObjectMapIterator();
   }
   {
     assert(NULL != index.index);
@@ -4985,7 +4985,7 @@ int FileStore::_create_collection(
   return 0;
 }
 
-int FileStore::_destroy_collection(coll_t c) 
+int FileStore::_destroy_collection(coll_t c)
 {
   int r = 0;
   char fn[PATH_MAX];
@@ -5030,7 +5030,7 @@ int FileStore::_collection_add(coll_t c, coll_t oldcid, const ghobject_t& o,
 			       const SequencerPosition& spos)
 {
   dout(15) << "collection_add " << c << "/" << o << " from " << oldcid << "/" << o << dendl;
-  
+
   int dstcmp = _check_replay_guard(c, o, spos);
   if (dstcmp < 0)
     return 0;
@@ -5335,7 +5335,7 @@ int FileStore::_split_collection(coll_t cid,
 
       assert(NULL != to.index);
       RWLock::WLocker l2((to.index)->access_lock);
-      
+
       r = from->split(rem, bits, to.index);
     }
