+++ src/test/filestore/TestFileStore.cc
@@ -16,7 +16,7 @@
 
 #include "common/ceph_argparse.h"
 #include "global/global_init.h"
-#include "os/FileStore.h"
+#include "os/filestore/FileStore.h"
 #include <gtest/gtest.h>
 
 class TestFileStore {
