+++ src/test/objectstore/FileStoreDiff.cc
@@ -15,7 +15,7 @@
 #include <map>
 #include <boost/scoped_ptr.hpp>
 #include "common/debug.h"
-#include "os/FileStore.h"
+#include "os/filestore/FileStore.h"
 #include "common/config.h"
 
 #include "FileStoreDiff.h"
