+++ src/test/objectstore/FileStoreTracker.h
@@ -3,7 +3,7 @@
 #ifndef FILESTORE_TRACKER_H
 #define FILESTORE_TRACKER_H
 #include "test/common/ObjectContents.h"
-#include "os/FileStore.h"
+#include "os/filestore/FileStore.h"
 #include "kv/KeyValueDB.h"
 #include <boost/scoped_ptr.hpp>
 #include <list>
