+++ src/test/objectstore/chain_xattr.cc
@@ -21,7 +21,7 @@
 
 #include <stdio.h>
 #include <signal.h>
-#include "os/chain_xattr.h"
+#include "os/filestore/chain_xattr.h"
 #include "include/Context.h"
 #include "common/errno.h"
 #include "common/ceph_argparse.h"
