+++ src/test/objectstore/store_test.cc
@@ -18,7 +18,7 @@
 #include <time.h>
 #include <sys/mount.h>
 #include "os/ObjectStore.h"
-#include "os/FileStore.h"
+#include "os/filestore/FileStore.h"
 #include "os/KeyValueStore.h"
 #include "include/Context.h"
 #include "common/ceph_argparse.h"
