+++ src/tools/ceph_objectstore_tool.cc
@@ -25,7 +25,7 @@
 #include "global/global_init.h"
 
 #include "os/ObjectStore.h"
-#include "os/FileJournal.h"
+#include "os/filestore/FileJournal.h"
 
 #include "osd/PGLog.h"
 #include "osd/OSD.h"
